[Parley] refactor: Split MainViewModel into partial classes (#536)
Split 1,786-line MainViewModel into 6 focused partial class files:
- Core (337 lines): Properties, constructor, debug messaging
- FileOperations (199): Load, Save, New, Close, Reload
- EditOperations (377): Undo, Redo, Copy, Cut, Paste
- NodeOperations (271): Add, Delete, Move nodes
- TreeOperations (593): Tree population, refresh, navigation
- ScrapOperations (91): Scrap restore and management

81% reduction in core file size. All 500 tests pass.
